requirements.txt
streamlit
pandas
mysql.connector
plotly.express